rename circuit_validator modules for unambiguous identity
Two files shared the name circuit_validator.py despite validating
different things. Rename to self-documenting names:

  models/circuit_validator.py        → models/circuit_schema_validator.py
  simulation/circuit_validator.py    → simulation/circuit_semantic_validator.py

Update all four call sites (file_controller, simulation/__init__,
grading/batch_grader, tests/unit/test_circuit_validator) and update
the test module docstring to match.

2772 tests pass; 1 pre-existing QSettings test failure unchanged.
