+++ app/simulation/__init__.py
@@ -1,5 +1,5 @@
 from . import circuitikz_exporter, convergence, csv_exporter
-from .circuit_validator import validate_circuit
+from .circuit_semantic_validator import validate_circuit
 from .netlist_generator import NetlistGenerator, generate_analysis_command
 from .ngspice_runner import NgspiceRunner
 from .result_parser import ResultParser
